+++ src/OpenTelemetry.Instrumentation.AspNetCore/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.13.0
+
+Released 2025-Oct-22
+
 * Added support for listening to ASP.NET Core Blazor activities.
   Configurable with the
   `AspNetCoreTraceInstrumentationOptions.EnableRazorComponentsSupport`
@@ -54,8 +58,8 @@
   page](https://learn.microsoft.com/dotnet/core/diagnostics/built-in-metrics-aspnetcore).
   ([#3012](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3012))
 
-* Updated OpenTelemetry core component version(s) to `1.13.0`.
-  ([#3158](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3158))
+* Updated OpenTelemetry core component version(s) to `1.13.1`.
+  ([#3218](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3218))
 
 ## 1.12.0