+++ src/OpenTelemetry.Resources.Gcp/CHANGELOG.md
@@ -17,5 +17,5 @@
 * Add support for FaaS resource attributes for Google Cloud Run.
   ([#2789](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2789))
 
-* Updated OpenTelemetry core component version(s) to `1.13.0`.
-  ([#3158](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3158))
+* Updated OpenTelemetry core component version(s) to `1.13.1`.
+  ([#3218](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3218))